requirements.txt
aiosmtplib==2.0.1
alembic==1.10.3
anyio==3.6.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
cloudinary==1.40.0
colorama==0.4.6
cryptography==40.0.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.3
fastapi==0.95.0
fastapi-camelcase==1.0.5
fastapi-mail==1.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
install==1.3.5
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pyhumps==3.8.0
pytest==7.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.40.0
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.6.0
urllib3==2.2.2
uvicorn==0.21.1
websockets==11.0.2
